requirements.txt
streamlit
psycopg2-binary
pandas
openpyxl
xlrd>=2.0.1
matplotlib
openai>=1.0.0,<2
argon2-cffi>=23.1.0
